requirements.txt
# THIVLVC dependencies
openai>=1.0.0             # OpenRouter + local (Ollama/vLLM) via OpenAI-compatible API
scikit-learn>=1.3.0       # TF-IDF retrieval
numpy>=1.24.0
requests>=2.31.0          # UDPipe/Lindat API
python-dotenv>=1.0.0
